requirements.txt
Django==5.1.2
djangorestframework==3.15.2
drf-spectacular==0.27.2